+++ test/jest.env.js
@@ -1,7 +1,9 @@
 Error.stackTraceLimit = 50;
 
-// const Rule = require('../src/core/Rule');
+const Validator = require('validator');
+const Rule = require('../src/core/Rule');
 
+Rule.factory('email', () => (f, v) => !Validator.isEmail(v), { enumerable: true });
 // Rule.extend('bookName', Rule.deny('The Bible'));
 // Rule.extend('bookPrice', Rule.range(0, 100));
 // Rule.extend('artComment', Rule.allow('yay', 'great', 'boo'));